Don't add reload_url in start-chrome-for-playr.sh
The playr_loader.html already (attempts to) add this parameter to the url.
Additionally, the current code in start-chrome-for-player.sh only
passes the reload_url parameter to playr_loader.html (which ignores
it), not the actual channel page on the server, so this parameter was
pointless.

Additionally, it seems impossible to add the reload_urlparameter here,
since if the reload_url parameter must be passed to playr_loader.html,
it must be present inside reload_url itself, which must contain another
reload_url option, etc.