Introduce keep_append option
This option works similar to the boot_append option, in that it causes
options to be appended to the commandline when generating a bootloader
configuration.

However, in addition to adding the contained option to the commandline,
the keep_append option itself is also added to the commandline again.
This means that when booting from the generated bootloader config, any
new bootloader configs generated will again have these options present
(in particular, the options inside keep_append will survive subsequent
upgrades).