Reinstall all packages
This was done by reinstalling each package separately, doing them all in one
command resulted in dependency loop errors.

	for p in $(dpkg -l |grep ^ii | cut -d' ' -f 3); do apt-get install --reinstall $p; done

This serves to check that nothing got missing or corrupted over time. A few
missing files were reinstalled by this, but nothing seems to have been really
broken.

The installer also updated the initrd, but this change was not committed since
there will be more changes for the initrd next.