Added fucntionality that will extend second partition on drive to enable persistence of user data
The second partition on the boot device will be enlarged to a max of 3GB
to store user data (browser cache and settings of Google Chrome).
The maximum of 3 GB was chosen because having no max and testing on 16GB
USB devices, even USB3, the resize took multiple minutes. This amount of
time will cause users to restart their devices before the resize is
finished which is too big of a risk of corruption.

Checks are in place to only resize the second partition of the device
that is used to boot from, if it is possible and needed.

The final act is mounting the second partition to /mnt/persistent which
location is used when starting up Google Chrome. Here the 'preparations'
of the scripts in the Makefile and install-to-disk.sh come together in a
unified location where data can be persisted that will survive system
reboots.