Added auto update functionality to Webconverger running on USB thumb drive, fixes in extending user data
Code written by Matthijs Kooijman was used as a basis.

Prerequisite for this fucntionality to work is the fact taht during the
make process of creating an img file enough room is created on the
partition that is used to run Webconverger from to allow for the growing
amount of information in the git repository that is used as a basis fro
the git-fs that Webconverger runs on.

Removed the noupgrade and noconfig command line options. The second to
enable staged roll out of updates to specific (numbers of) users using
the ability for remote configuration of Webconverger..

a few changes were made to extend-user-data to make it work on both the
second partition of an USB thumb drive as well as on the second data
partition of the hard drive that Webconverger is installed on.
Changes were; additional debug info and more robust checks on whether resize is
needed and if execution of the logic can be done savely and changing
mount to use the file system type that is present. Also chown is used
instead of chmod to set the correct access rights on 1000 insted of webc
to make all related code do the same.

In install-to-disk the chwon was applied to the /mnt/persistent
directory and not only the persistent directory within it to make it
more in line with the logic for USB thumb drives